refactor: bulk edit action set changeTo as default value
