+++ packages/core/client/src/block-provider/hooks/index.ts
@@ -15,7 +15,6 @@ import { useFilterBlock } from '../../filter-provider/FilterProvider';
 import { transformToFilter } from '../../filter-provider/utils';
 import { useRecord } from '../../record-provider';
 import { removeNullCondition, useActionContext, useCompile } from '../../schema-component';
-import { BulkEditFormItemValueType } from '../../schema-initializer/components';
 import { useCurrentUserContext } from '../../user';
 import { useLocalVariables, useVariables } from '../../variables';
 import { isVariable } from '../../variables/utils/isVariable';
@@ -676,94 +675,6 @@ export const useCustomizeBulkUpdateActionProps = () => {
           actionField.data.loading = false;
         },
       });
-    },
-  };
-};
-
-export const useCustomizeBulkEditActionProps = () => {
-  const form = useForm();
-  const { t } = useTranslation();
-  const { field, resource, __parent } = useBlockRequestContext();
-  const expressionScope = useContext(SchemaExpressionScopeContext);
-  const actionContext = useActionContext();
-  const navigate = useNavigate();
-  const compile = useCompile();
-  const actionField = useField();
-  const tableBlockContext = useTableBlockContext();
-  const { modal } = App.useApp();
-
-  const { rowKey } = tableBlockContext;
-  const selectedRecordKeys =
-    tableBlockContext.field?.data?.selectedRowKeys ?? expressionScope?.selectedRecordKeys ?? {};
-  const { setVisible, fieldSchema: actionSchema } = actionContext;
-  return {
-    async onClick() {
-      const { onSuccess, skipValidator, updateMode } = actionSchema?.['x-action-settings'] ?? {};
-      const { filter } = __parent.service.params?.[0] ?? {};
-      if (!skipValidator) {
-        await form.submit();
-      }
-      const values = cloneDeep(form.values);
-      actionField.data = field.data || {};
-      actionField.data.loading = true;
-      for (const key in values) {
-        if (Object.prototype.hasOwnProperty.call(values, key)) {
-          const value = values[key];
-          if (BulkEditFormItemValueType.Clear in value) {
-            values[key] = null;
-          } else if (BulkEditFormItemValueType.ChangedTo in value) {
-            values[key] = value[BulkEditFormItemValueType.ChangedTo];
-          } else if (BulkEditFormItemValueType.RemainsTheSame in value) {
-            delete values[key];
-          }
-        }
-      }
-      try {
-        const updateData: { filter?: any; values: any; forceUpdate: boolean } = {
-          values,
-          filter,
-          forceUpdate: false,
-        };
-        if (updateMode === 'selected') {
-          if (!selectedRecordKeys?.length) {
-            message.error(t('Please select the records to be updated'));
-            return;
-          }
-          updateData.filter = { $and: [{ [rowKey || 'id']: { $in: selectedRecordKeys } }] };
-        }
-        if (!updateData.filter) {
-          updateData.forceUpdate = true;
-        }
-        await resource.update(updateData);
-        actionField.data.loading = false;
-        if (!(resource instanceof TableFieldResource)) {
-          __parent?.__parent?.service?.refresh?.();
-        }
-        __parent?.service?.refresh?.();
-        setVisible?.(false);
-        if (!onSuccess?.successMessage) {
-          return;
-        }
-        if (onSuccess?.manualClose) {
-          modal.success({
-            title: compile(onSuccess?.successMessage),
-            onOk: async () => {
-              await form.reset();
-              if (onSuccess?.redirecting && onSuccess?.redirectTo) {
-                if (isURL(onSuccess.redirectTo)) {
-                  window.location.href = onSuccess.redirectTo;
-                } else {
-                  navigate(onSuccess.redirectTo);
-                }
-              }
-            },
-          });
-        } else {
-          message.success(compile(onSuccess?.successMessage));
-        }
-      } finally {
-        actionField.data.loading = false;
-      }
     },
   };
 };
