+++ packages/core/client/src/schema-initializer/components/index.ts
@@ -1,3 +1,2 @@
 export * from './assigned-field';
-export * from './BulkEditField';
 export * from './CreateRecordAction';
