+++ packages/core/client/src/schema-initializer/index.ts
@@ -1,7 +1,6 @@
 import { Plugin } from '../application/Plugin';
 import {
   blockInitializers,
-  bulkEditFormActionInitializers,
   createFormActionInitializers,
   createFormBlockInitializers,
   cusomeizeCreateFormBlockInitializers,
@@ -74,7 +73,6 @@ export class SchemaInitializerPlugin extends Plugin {
     this.app.schemaInitializerManager.add(filterFormActionInitializers);
     this.app.schemaInitializerManager.add(createFormActionInitializers);
     this.app.schemaInitializerManager.add(updateFormActionInitializers);
-    this.app.schemaInitializerManager.add(bulkEditFormActionInitializers);
     this.app.schemaInitializerManager.add(filterFormItemInitializers);
     this.app.schemaInitializerManager.add(gridCardActionInitializers);
     this.app.schemaInitializerManager.add(gridCardItemActionInitializers);
